+++ sdks/python/hatchet_sdk/context/context.py
@@ -25,12 +25,8 @@ from hatchet_sdk.exceptions import TaskRunError
 from hatchet_sdk.features.runs import RunsClient
 from hatchet_sdk.logger import logger
 from hatchet_sdk.utils.timedelta_to_expression import Duration, timedelta_to_expr
-from hatchet_sdk.utils.typing import JSONSerializableMapping
-from hatchet_sdk.worker.runner.utils.capture_logs import (
-    AsyncLogSender,
-    LogLevel,
-    LogRecord,
-)
+from hatchet_sdk.utils.typing import JSONSerializableMapping, LogLevel
+from hatchet_sdk.worker.runner.utils.capture_logs import AsyncLogSender, LogRecord
 
 if TYPE_CHECKING:
     from hatchet_sdk.runnables.task import Task
